Review comments + reformatting
